Create LCM_calculation.c
C Program to Find LCM of two Numbers

To understand this example, you should have the knowledge of the following C programming topics:

    #C Programming Operators
    #C if...else Statement
    #C while and do...while Loop

The LCM of two integers n1 and n2 is the smallest positive integer that is perfectly divisible by both n1 and n2 (without a remainder). For example, the LCM of 72 and 120 is 360.